agent_workflow.py
from typing import Dict, List, Tuple, TypedDict, Annotated
from langgraph.graph import Graph, StateGraph
from langgraph.prebuilt import ToolNode
from pydantic import BaseModel, Field
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# State definitions
class AgentState(TypedDict):
    user_prompt: str
    rag_documents: List[str]
    draft_scripts: List[str]
    consolidated_script: str
    verification_feedback: str
    iteration_count: int
    is_verified: bool

# Agent definitions
class DraftingAgent:
    def __init__(self, model_provider: str):
        self.model_provider = model_provider
        # Placeholder for model initialization
        pass

    def draft_script(self, state: AgentState) -> AgentState:
        # Placeholder for RAG document fetching
        rag_docs = state["rag_documents"]
        
        # Placeholder for script generation
        draft = f"Generated script using {self.model_provider}"
        
        state["draft_scripts"].append(draft)
        return state

class ConsolidationAgent:
    def __init__(self, model_provider: str):
        self.model_provider = model_provider
        pass

    def consolidate_scripts(self, state: AgentState) -> AgentState:
        # Placeholder for script consolidation
        consolidated = "Consolidated script"
        state["consolidated_script"] = consolidated
        return state

class VerificationAgent:
    def __init__(self, model_provider: str):
        self.model_provider = model_provider
        pass

    def verify_script(self, state: AgentState) -> AgentState:
        # Placeholder for verification logic
        is_valid = True  # Placeholder
        feedback = "Verification feedback"
        
        state["verification_feedback"] = feedback
        state["is_verified"] = is_valid
        return state

class CorrectionsAgent:
    def __init__(self, model_provider: str):
        self.model_provider = model_provider
        pass

    def correct_script(self, state: AgentState) -> AgentState:
        # Placeholder for correction logic
        corrected_script = "Corrected script"
        state["consolidated_script"] = corrected_script
        state["iteration_count"] += 1
        return state

def should_continue_verification(state: AgentState) -> str:
    if state["is_verified"]:
        return "end"
    if state["iteration_count"] >= 5:
        return "end"
    return "continue_verification"

def create_workflow() -> Graph:
    # Initialize agents
    drafting_agent1 = DraftingAgent("provider1")
    drafting_agent2 = DraftingAgent("provider2")
    drafting_agent3 = DraftingAgent("provider3")
    consolidation_agent = ConsolidationAgent("provider1")
    verification_agent = VerificationAgent("provider1")
    corrections_agent = CorrectionsAgent("provider1")

    # Create workflow
    workflow = StateGraph(AgentState)

    # Add nodes
    workflow.add_node("drafting1", drafting_agent1.draft_script)
    workflow.add_node("drafting2", drafting_agent2.draft_script)
    workflow.add_node("drafting3", drafting_agent3.draft_script)
    workflow.add_node("consolidation", consolidation_agent.consolidate_scripts)
    workflow.add_node("verification", verification_agent.verify_script)
    workflow.add_node("corrections", corrections_agent.correct_script)

    # Add edges
    workflow.add_edge("drafting1", "drafting2")
    workflow.add_edge("drafting2", "drafting3")
    workflow.add_edge("drafting3", "consolidation")
    workflow.add_edge("consolidation", "verification")
    workflow.add_edge("verification", "corrections")
    workflow.add_edge("corrections", "verification")

    # Set entry and conditional edges
    workflow.set_entry_point("drafting1")
    workflow.add_conditional_edges(
        "verification",
        should_continue_verification,
        {
            "continue_verification": "corrections",
            "end": "end"
        }
    )

    return workflow.compile()

def main():
    # Initialize workflow
    workflow = create_workflow()
    
    # Example usage
    initial_state = {
        "user_prompt": "Example prompt",
        "rag_documents": [],
        "draft_scripts": [],
        "consolidated_script": "",
        "verification_feedback": "",
        "iteration_count": 0,
        "is_verified": False
    }
    
    # Run workflow
    result = workflow.invoke(initial_state)
    print("Final state:", result)

if __name__ == "__main__":
    main() 